+++ domain/src/main/java/com/terning/domain/entity/HomeFilteringInfoModel.kt
@@ -1,4 +1,4 @@
-package com.terning.domain.entity.response
+package com.terning.domain.entity
 
 data class HomeFilteringInfoModel(
     val grade: Int?,
